# [66_5] Perf tuning on tm_reader::decode via index_of