--- a/src/Data/Convert/Texmacs/fromtm.cpp
+++ b/src/Data/Convert/Texmacs/fromtm.cpp
@@ -80,8 +80,8 @@ tm_reader<format_without_utf8>::decode (string s) {
   if (start == -1) return s;
 
   int    n= N (s);
-  string r;
-  for (int i= 0; i < n; i++)
+  string r= s(0, start);
+  for (int i= start; i < n; i++)
     if (((i + 1) < n) && (s[i] == '\\')) {
       i++;
       if (s[i] == ';')