--- a/src/Data/Convert/Texmacs/fromtm.cpp
+++ b/src/Data/Convert/Texmacs/fromtm.cpp
@@ -79,9 +79,9 @@ tm_reader<format_without_utf8>::decode (string s) {
   int start= index_of (s, '\\');
   if (start == -1) return s;
 
-  int    n= N (s);
+  int    n_minus_one= N (s) - 1, i= start;
   string r= s (0, start);
-  for (int i= start; i < n - 1; i++)
+  for (; i < n_minus_one; i++)
     if (s[i] == '\\') {
       i++;
       if (s[i] == ';')