@darcy-shen 2024-02-16
Tested on Linux, there is no significant performance improvements.