@jingkaimori 2024-02-16
Performance is slightly improved:

Before:
```
PASS   : TestConverter::test_texmacs_to_tree_bench(29_1_1.tm)
RESULT : TestConverter::test_texmacs_to_tree_bench():"29_1_1.tm":
     0.018 msecs per iteration (total: 76, iterations: 4096)
PASS   : TestConverter::test_texmacs_to_tree_bench(46_3.tm)
RESULT : TestConverter::test_texmacs_to_tree_bench():"46_3.tm":
     0.394 msecs per iteration (total: 101, iterations: 256)
PASS   : TestConverter::test_texmacs_to_tree_bench(64_1.tm)
RESULT : TestConverter::test_texmacs_to_tree_bench():"64_1.tm":
     1.5 msecs per iteration (total: 98, iterations: 64)
```

After:
```
PASS   : TestConverter::test_texmacs_to_tree_bench(29_1_1.tm)
RESULT : TestConverter::test_texmacs_to_tree_bench():"29_1_1.tm":
     0.016 msecs per iteration (total: 69, iterations: 4096)
PASS   : TestConverter::test_texmacs_to_tree_bench(46_3.tm)
RESULT : TestConverter::test_texmacs_to_tree_bench():"46_3.tm":
     0.37 msecs per iteration (total: 95, iterations: 256)
PASS   : TestConverter::test_texmacs_to_tree_bench(64_1.tm)
RESULT : TestConverter::test_texmacs_to_tree_bench():"64_1.tm":
     1.4 msecs per iteration (total: 92, iterations: 64)
```